Differentiate scan and performance code
Performance behavior can be triggered by specifying a period time of 0.
In any other case, the code performs a proper scan cycles at the given
period.

In scan mode, the speed of the EtherCAT packet is masked by the period:

    1. Receive the packet from EtherCAT (inputs are updated)
    2. Execute the cycle
    3. Send a new packet to EtherCAT (outputs are updated)
    4. Show the span time, i.e. the time needed to execute 1, 2 and 3
    5. Wait the next period
    6. Repeat

In performance mode, there is no cycle to be executed and the EtherCAT
speed is measured directly:

    1. Send a new packet to EtherCAT
    2. Receive the packet from EtherCAT
    3. Show the roundtrip time, i.e. the time needed to execute 1 and 2
    4. Repeat